+++ CHANGELOG.md
@@ -38,3 +38,4 @@ Created on Dec 20, 2023
 - Beta Release: Navbar bugs fixed on March 10, 2024 -> **0.625**
 - Beta Release: New implementation on lookup transactions, overall improvements, and re organized edit transaction on March 10, 2024 -> **0.7**
 - Beta Release: Overall improvements, and draft transaction modification on March 10, 2024 -> **0.72**
+- Beta Release: Transaction modification is functional and overall improvements on March 11, 2024 -> **0.75**